Comment: (github.styl) Issue label text color
Black provides poor contrast with some dark label colors, so using
white seems like the best choice for maximum contrast with arbitrary
label colors.  But since the default label color is light gray, that
would make for very poor contrast on new labels.  So we use a CSS drop
shadow to compensate for that.  Hopefully it's a reasonable
compromise, because there's only so much that can be done.

But none of this works because Stylus can't handle :not(), or
something like that.  See comment in file.  So this is really just a
note for the future, because I don't feel like hacking up all the
generated CSS for the sake of GitHub labels.  *sigh*